fix some assertions on 64-bit osx builds
git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/trunk@953 69509848-2a73-49d5-a893-c4d8c934025a

+++ src/grandorgue/GrandOrgueFile.cpp
@@ -193,9 +193,10 @@ bool GrandOrgueFile::TryLoad
 			}
 			GOrgueLCD_WriteLineTwo
 				(wxString::Format
-				 (_("Loading %d%%")
-				  ,(nb_loaded_obj * 100) / (objects.size() + 1))
-				 );
+					(_("Loading %lu%%")
+					,(unsigned long)((nb_loaded_obj * 100) / (objects.size() + 1))
+					)
+				);
 		}
 
 		success = true;
